refactor(content): move related and discovery out; delete markdown.ts
The decomposition is complete: getRelatedPosts/getAdjacentPosts live in
content/related.ts, and the cross-content aggregates (getAllTags, the
wikilink slug registry, the backlink index) live in content/discovery.ts —
they span posts+flows+notes+series, which is why they sit at the top of
the dependency chain. The prod-only registry/backlink singletons become
createProdMemo (same dev-rebuild semantics). markdown.ts is deleted;
remark-wikilinks' SlugRegistryEntry type import moves to discovery.

New tests/integration/discovery.test.ts pins previously untested behavior:
bare-slug registry keying (collisions resolve last-write-wins — pinned,
not papered over), note aliases, backlink extraction with context bounds,
self-reference exclusion, and tag-casing dedupe.

Docs updated in the same change: CLAUDE.md integration-point/turbopackIgnore/
verify rules now point at src/lib/content/**; ARCHITECTURE.md gains the
content/ module map with the dependency diagram; AGENTS.md/GEMINI.md refs
updated.
