+++ src/app/tags/[tag]/page.tsx
@@ -1,4 +1,4 @@
-import { getAllTags } from '@/lib/markdown';
+import { getAllTags } from '@/lib/content/discovery';
 import { getPostsByTag } from '@/lib/content/posts';
 import { getFlowsByTag } from '@/lib/content/flows';
 import { notFound } from 'next/navigation';
