+++ src/layouts/PostLayout.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { Suspense } from 'react';
-import { BacklinkSource, SlugRegistryEntry } from '@/lib/markdown';
+import type { BacklinkSource, SlugRegistryEntry } from '@/lib/content/discovery';
 import { getAuthorSlug } from '@/lib/content/authors';
 import type { PostData, CollectionContext } from '@/lib/content/types';
 import MarkdownRenderer from '@/components/MarkdownRenderer';
